+++ pkg/oas/oas_response_encoders_gen.go
@@ -461,6 +461,20 @@ func encodeGetBlockchainConfigResponse(response *BlockchainConfig, w http.Respon
 	return nil
 }
 
+func encodeGetBlockchainConfigFromBlockResponse(response *BlockchainConfig, w http.ResponseWriter, span trace.Span) error {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(200)
+	span.SetStatus(codes.Ok, http.StatusText(200))
+
+	e := jx.GetEncoder()
+	response.Encode(e)
+	if _, err := e.WriteTo(w); err != nil {
+		return errors.Wrap(err, "write")
+	}
+
+	return nil
+}
+
 func encodeGetBlockchainMasterchainHeadResponse(response *BlockchainBlock, w http.ResponseWriter, span trace.Span) error {
 	w.Header().Set("Content-Type", "application/json")
 	w.WriteHeader(200)