+++ pkg/oas/oas_router_gen.go
@@ -720,18 +720,17 @@ func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 							break
 						}
 						switch elem[0] {
-						case 'c': // Prefix: "config/raw"
-							if l := len("config/raw"); len(elem) >= l && elem[0:l] == "config/raw" {
+						case 'c': // Prefix: "config"
+							if l := len("config"); len(elem) >= l && elem[0:l] == "config" {
 								elem = elem[l:]
 							} else {
 								break
 							}
 
 							if len(elem) == 0 {
-								// Leaf node.
 								switch r.Method {
 								case "GET":
-									s.handleGetRawBlockchainConfigFromBlockRequest([1]string{
+									s.handleGetBlockchainConfigFromBlockRequest([1]string{
 										args[0],
 									}, elemIsEscaped, w, r)
 								default:
@@ -739,6 +738,28 @@ func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 								}
 
 								return
+							}
+							switch elem[0] {
+							case '/': // Prefix: "/raw"
+								if l := len("/raw"); len(elem) >= l && elem[0:l] == "/raw" {
+									elem = elem[l:]
+								} else {
+									break
+								}
+
+								if len(elem) == 0 {
+									// Leaf node.
+									switch r.Method {
+									case "GET":
+										s.handleGetRawBlockchainConfigFromBlockRequest([1]string{
+											args[0],
+										}, elemIsEscaped, w, r)
+									default:
+										s.notAllowed(w, r, "GET")
+									}
+
+									return
+								}
 							}
 						case 't': // Prefix: "transactions"
 							if l := len("transactions"); len(elem) >= l && elem[0:l] == "transactions" {
@@ -3064,8 +3085,8 @@ func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
 							break
 						}
 						switch elem[0] {
-						case 'c': // Prefix: "config/raw"
-							if l := len("config/raw"); len(elem) >= l && elem[0:l] == "config/raw" {
+						case 'c': // Prefix: "config"
+							if l := len("config"); len(elem) >= l && elem[0:l] == "config" {
 								elem = elem[l:]
 							} else {
 								break
@@ -3074,16 +3095,39 @@ func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
 							if len(elem) == 0 {
 								switch method {
 								case "GET":
-									// Leaf: GetRawBlockchainConfigFromBlock
-									r.name = "GetRawBlockchainConfigFromBlock"
+									r.name = "GetBlockchainConfigFromBlock"
 									r.summary = ""
-									r.operationID = "getRawBlockchainConfigFromBlock"
-									r.pathPattern = "/v2/blockchain/blocks/{block_id}/config/raw"
+									r.operationID = "getBlockchainConfigFromBlock"
+									r.pathPattern = "/v2/blockchain/blocks/{block_id}/config"
 									r.args = args
 									r.count = 1
 									return r, true
 								default:
 									return
+								}
+							}
+							switch elem[0] {
+							case '/': // Prefix: "/raw"
+								if l := len("/raw"); len(elem) >= l && elem[0:l] == "/raw" {
+									elem = elem[l:]
+								} else {
+									break
+								}
+
+								if len(elem) == 0 {
+									switch method {
+									case "GET":
+										// Leaf: GetRawBlockchainConfigFromBlock
+										r.name = "GetRawBlockchainConfigFromBlock"
+										r.summary = ""
+										r.operationID = "getRawBlockchainConfigFromBlock"
+										r.pathPattern = "/v2/blockchain/blocks/{block_id}/config/raw"
+										r.args = args
+										r.count = 1
+										return r, true
+									default:
+										return
+									}
 								}
 							}
 						case 't': // Prefix: "transactions"